requirements.txt
imutils~=0.5.4
opencv-python~=4.11.0.86
# requests~=2.32.3
Flask~=3.1.0
scikit-image~=0.25.2
pillow~=11.1.0
gunicorn